backend/requirements.txt
fastapi
uvicorn
pymupdf
python-multipart
spacy
google-genai
python-dotenv
